Add paparazzi for snapshot testing and some needed changes to support it
