+++ debugview/src/debug/kotlin/com/revenuecat/purchases/debugview/settings/SettingGroup.kt
@@ -13,11 +13,13 @@ import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.revenuecat.purchases.Package
 import com.revenuecat.purchases.debugview.DebugRevenueCatViewModel
+import com.revenuecat.purchases.debugview.findActivity
 import com.revenuecat.purchases.debugview.models.SettingGroupState
 import com.revenuecat.purchases.debugview.models.SettingScreenState
 import com.revenuecat.purchases.debugview.models.SettingState
@@ -29,6 +31,7 @@ import kotlinx.coroutines.flow.StateFlow
 internal fun SettingGroup(
     settingGroupState: SettingGroupState,
     viewModel: DebugRevenueCatViewModel,
+    activity: Activity = LocalContext.current.findActivity(),
 ) {
     Column(
         modifier = Modifier
@@ -49,7 +52,11 @@ internal fun SettingGroup(
                 settingGroupState.settings.forEach { settingState ->
                     when (settingState) {
                         is SettingState.Text -> SettingText(settingState)
-                        is SettingState.OfferingSetting -> SettingOffering(settingState, screenViewModel = viewModel)
+                        is SettingState.OfferingSetting -> SettingOffering(
+                            settingState,
+                            activity = activity,
+                            screenViewModel = viewModel,
+                        )
                     }
                     Divider()
                 }
@@ -58,7 +65,7 @@ internal fun SettingGroup(
     }
 }
 
-@Preview
+@Preview(showBackground = true)
 @Composable
 private fun SettingGroupPreview() {
     val viewModel = object : DebugRevenueCatViewModel {
