+++ gradle.properties
@@ -31,6 +31,8 @@ POM_DEVELOPER_NAME=RevenueCat, Inc.
 
 android.useAndroidX=true
 android.enableJetifier=true
+# This is needed for paparazzi to work through jetifier.
+android.jetifier.ignorelist=android-base-common,common
 
 # Avoid timeouts when submitting to Sonatype
 systemProp.org.gradle.internal.http.socketTimeout=120000
